fix: satisfy camera handler lint rules
